Started the Rotation
Currently rotating everything to be worked with easier. The entire level and the enemy movements will all need to be edited to move in an xy-plane instead of a xz-plane. Also will need to program some friction so that objects don't move more than they should (look up formula for doing so!)
I also have the beginnings of a map navigation, which is awesome!