Data Recovered!
Some fixes have been made to the nodemap maker to make it require less data and run faster. Currently working on an issue where if a player or enemy has a node location inside of them, then it won't have any friends, since the player collider blocks it. That, and making sure the findNearestNode returns a node WITH FRIENDS! That is needed!