Queens Path Works!
Some Files were removed because of duplication, and with a test, we confirmed that the Queen's path sets properly upon startup. Next, we'll see what happens as we have our wolves update it if they're at the end of the queen's path and the player isn't there. This'll be great!