Finished Cleaning/Queens Path Works!
The Queen's path works alright. It does what I want it to do, which is good, but there's one problem: The lag is still there! We will need a new means to fix it... Perhaps allowing the Queens path to form without a while loop, but with some recursion to allow for multitasking? It's worth a shot, at least! We will see how it works.
Also, we got rid of some of the Github files on this drive that were just taking up space.