Can Now Visualize Map!
Created visual markers to better see how the map looks and generates. It seems that the map works fine, but the generation of the nodes still causes some to be spawned inside walls, even though the map was designed to be an 8x8 grid... Overall the map works, except the monsters keep finding places to get stuck in. This can prove to be a problem in the future, and even so, the original tracePath function remains more accurate.
Considering placing nodes manually, but tracePath works just fine without them. Will study the function to see what that one does right. Will also need to study the dogs to see why they keep getting stuck with the cachemap function.