NodeMap Forms Correctly!
The current nodemap is now formed to have the right number of nodes to fit it's map (turns out, this is a 12x12 map, not a 16^2 map). Now, the next issue is that some dogs get a path in reverse from how they need it. This can be fixed in the dog's code, but may be more effectively done in the nodemap code (have the cache checker check which end of the map is closest to the start, and have it return the map starting with that starting position in the right place, instead of having all the other enemies have to adjust themselves over and over).