NodeMap is Good Enough!
The map works well enough that it will do what we need it to do for now. (And with the town, we don't even need it, so we'll be fine without it). Also, added the cool feature that all the children stare at Reu while he's walking around (not quite as creepy as I had hoped, but it does it's job).
Next, we have a few options for a next step:
1. Finishing the wolves' programming (for testing if this game will be fun.
2. Work on Scarlett's design and programming!
3. Make the camera never show what is out of bounds!
4. See if we can start working on a monster-trapping method. If not, we may want the wolves to be unhurtable and/or have Reu collect things other than animal souls, such as.. teddy bears? Something Scarlett needs for sustenance, like finding kids who get lost and bringing them home! (maybe in the coffin! That would be horrific and genius!)

Overall, small progress is going in the right direction!